scripts/requirements.txt
pykwalify==1.8.0
pyyaml==6.0.1
yamllint==1.35.1
httpx==0.27.0
pybatfish==2024.4.16
jinja2==3.1.4
